+++ src/configuration/WithTemplateConfig/constants.tsx
@@ -1,0 +1,9 @@
+import { IconName } from "@grafana/ui";
+
+export const generalFolderTitle = "General"
+
+export const infoTab = {
+  id: "info",
+  label: "How it works",
+  icon: "info-circle" as IconName,
+}