+++ src/querybuilder/components/PromQueryEditorSelector.tsx
@@ -16,7 +16,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import React, { SyntheticEvent, useCallback, useEffect, useState } from 'react';
+import React, { SyntheticEvent, useCallback, useEffect, useMemo, useState } from 'react';
 
 import { CoreApp, LoadingState } from '@grafana/data';
 import { reportInteraction } from '@grafana/runtime';
@@ -25,6 +25,9 @@ import { Button, ConfirmModal } from '@grafana/ui';
 import { EditorHeader, EditorRows, FlexItem, InlineSelect, Space } from '../../components/QueryEditor';
 import VmuiLink from "../../components/VmuiLink";
 import { PromQueryEditorProps } from '../../components/types';
+import WithTemplatePreview from "../../configuration/WithTemplateConfig/WithTemplatesPreview/WithTemplatePreview";
+import { getArrayFromTemplate } from "../../configuration/WithTemplateConfig/utils/getArrayFromTemplate";
+import PrometheusLanguageProvider from "../../language_provider";
 import { PromQuery } from '../../types';
 import { promQueryModeller } from '../PromQueryModeller';
 import { buildVisualQueryFromString } from '../parsing';
@@ -37,15 +40,26 @@ import { changeEditorMode, getQueryWithDefaults } from '../state';
 import { PromQueryBuilderContainer } from './PromQueryBuilderContainer';
 import { PromQueryBuilderOptions } from './PromQueryBuilderOptions';
 import { PromQueryCodeEditor } from './PromQueryCodeEditor';
+import { QueryPreview } from "./QueryPreview";
+import { TraceView } from "./Trace";
 
 type Props = PromQueryEditorProps;
 
 export const PromQueryEditorSelector = React.memo<Props>((props) => {
+
   const { onChange, onRunQuery, data, app, datasource } = props;
+
   const [parseModalOpen, setParseModalOpen] = useState(false);
   const [dataIsStale, setDataIsStale] = useState(false);
   const [trace, setTrace] = useState(false);
+  const [rawQuery, setRawQuery] = useState(false)
   const { flag: explain, setFlag: setExplain } = useFlag(promQueryEditorExplainKey);
+
+  const withTemplate = useMemo(() => {
+    // @ts-ignore
+    const dashboardUID = data?.request?.dashboardUID
+    return datasource.withTemplates.find(t => t.uid === dashboardUID)
+  }, [data, datasource])
 
   const query = getQueryWithDefaults(props.query, app);
   // This should be filled in from the defaults by now.
@@ -92,6 +106,15 @@ export const PromQueryEditorSelector = React.memo<Props>((props) => {
     onRunQuery();
   };
 
+  const onShowRawChange = (e: SyntheticEvent<HTMLInputElement>) => {
+    setRawQuery(e.currentTarget.checked)
+  }
+
+  useEffect(() => {
+    const withTemplates = getArrayFromTemplate(withTemplate)
+    datasource.languageProvider = new PrometheusLanguageProvider(datasource, { withTemplates })
+  }, [datasource, withTemplate])
+
   return (
     <>
       <ConfirmModal
@@ -122,10 +145,11 @@ export const PromQueryEditorSelector = React.memo<Props>((props) => {
           }}
           options={promQueryModeller.getQueryPatterns().map((x) => ({ label: x.name, value: x }))}
         />
-
         <QueryHeaderSwitch label="Explain" value={explain} onChange={onShowExplainChange}/>
         <QueryHeaderSwitch label="Trace" value={trace} onChange={onShowTracingChange}/>
+        <QueryHeaderSwitch label="Raw" value={rawQuery} onChange={onShowRawChange}/>
         <FlexItem grow={1}/>
+        {withTemplate && <WithTemplatePreview value={withTemplate.expr} datasource={datasource}/>}
         {app !== CoreApp.Explore && (
           <Button
             variant={dataIsStale ? 'primary' : 'secondary'}
@@ -143,12 +167,15 @@ export const PromQueryEditorSelector = React.memo<Props>((props) => {
       <Space v={0.5}/>
       <EditorRows>
         {editorMode === QueryEditorMode.Code && (
-          <PromQueryCodeEditor
-            {...props}
-            query={query}
-            showExplain={explain}
-            showTrace={trace}
-          />
+           <>
+             <PromQueryCodeEditor
+               {...props}
+               key={withTemplate?.uid}
+               query={query}
+               showExplain={explain}
+             />
+             {rawQuery && <QueryPreview query={query.expr} withTemplate={withTemplate} />}
+           </>
         )}
         {editorMode === QueryEditorMode.Builder && (
           <PromQueryBuilderContainer
@@ -157,11 +184,10 @@ export const PromQueryEditorSelector = React.memo<Props>((props) => {
             onChange={onChangeInternal}
             onRunQuery={props.onRunQuery}
             data={data}
-            showRawQuery={true}
             showExplain={explain}
-            showTrace={trace}
           />
         )}
+        {trace && <TraceView query={query} datasource={datasource} data={data} />}
         <PromQueryBuilderOptions query={query} app={props.app} onChange={onChange} onRunQuery={onRunQuery}/>
       </EditorRows>
     </>