+++ src/querybuilder/components/QueryPreview.tsx
@@ -19,19 +19,25 @@
 import React from 'react';
 
 import { EditorRow, EditorFieldGroup, EditorField } from '../../components/QueryEditor';
+import { WithTemplate } from "../../configuration/WithTemplateConfig/types";
+import { mergeTemplateWithQuery } from "../../configuration/WithTemplateConfig/utils/getArrayFromTemplate";
 import metricsqlGrammar from '../../metricsql';
 import { RawQuery } from '../shared/RawQuery';
 
 export interface Props {
   query: string;
+  withTemplate?: WithTemplate
 }
 
-export function QueryPreview({ query }: Props) {
+export function QueryPreview({ query, withTemplate }: Props) {
   return (
     <EditorRow>
       <EditorFieldGroup>
         <EditorField label="Raw query">
-          <RawQuery query={query} lang={{ grammar: metricsqlGrammar, name: 'promql' }} />
+          <RawQuery
+            query={mergeTemplateWithQuery(query, withTemplate)}
+            lang={{ grammar: metricsqlGrammar, name: 'promql' }}
+          />
         </EditorField>
       </EditorFieldGroup>
     </EditorRow>